some warnings fixed
git-svn-id: http://miranda.googlecode.com/svn/trunk/miranda/plugins/Scriver@3939 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ msgdialog.c
@@ -1381,7 +1381,7 @@ BOOL CALLBACK DlgProcMessage(HWND hwndDlg, UINT msg, WPARAM wParam, LPARAM lPara
 	case DM_GETCONTEXTMENU:
 		{
 			HMENU hMenu = (HMENU) CallService(MS_CLIST_MENUBUILDCONTACT, (WPARAM) dat->hContact, 0);
-			SetWindowLong(hwndDlg, DWL_MSGRESULT, hMenu);
+			SetWindowLong(hwndDlg, DWL_MSGRESULT, (LONG)hMenu);
 			return TRUE;
 		}
 	case WM_CONTEXTMENU:
